# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WINR platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: listMyWithdrawals
      tags: [withdrawals]
      summary: List my withdrawal requests
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Withdrawal" }
    post:
      operationId: requestWithdrawal
      tags: [withdrawals]
      summary: Request a withdrawal (1 per week)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestWithdrawalBody" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Withdrawal" }
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
  /admin/users/{userId}/balance:
    post:
      operationId: adminUpdateBalance
      tags: [admin]
      summary: Credit or debit a user's balance
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUpdateBalanceBody" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
  /admin/users/{userId}/status:
    post:
      operationId: adminSetUserStatus
      tags: [admin]
      summary: Activate, suspend, or reset a user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSetStatusBody" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
  /admin/payments:
    get:
      operationId: adminListPayments
      tags: [admin]
      summary: List payment claims
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminPayment" }
  /admin/payments/{paymentId}/decision:
    post:
      operationId: adminDecidePayment
      tags: [admin]
      summary: Confirm or reject a payment claim
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminDecidePaymentBody" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminPayment" }
  /admin/withdrawals:
    get:
      operationId: adminListWithdrawals
      tags: [admin]
      summary: List withdrawal requests
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminWithdrawal" }
  /admin/withdrawals/{withdrawalId}/decision:
    post:
      operationId: adminDecideWithdrawal
      tags: [admin]
      summary: Approve or reject a withdrawal
      parameters:
        - in: path
          name: withdrawalId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminDecideWithdrawalBody" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminWithdrawal" }
...
  /admin/affiliate/payouts/{referralId}/payout:
    post:
      operationId: adminPayoutAffiliate
      tags: [admin]
      summary: Mark a referral bonus as paid (credits the referrer)
      parameters:
        - in: path
          name: referralId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AffiliatePayout" }
  /admin/tickets:
    get:
      operationId: adminListTickets
      tags: [admin]
      summary: List support tickets
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SupportTicket" }
  /admin/tickets/{ticketId}:
    get:
      operationId: adminGetTicket
      tags: [admin]
      summary: Get ticket with messages (admin)
      parameters:
        - in: path
          name: ticketId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupportTicketDetail" }
  /admin/tickets/{ticketId}/reply:
    post:
      operationId: adminReplyTicket
      tags: [admin]
      summary: Admin reply to a ticket
      parameters:
        - in: path
          name: ticketId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostTicketMessageBody" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupportMessage" }
  /admin/notifications/broadcast:
    post:
      operationId: adminBroadcast
      tags: [admin]
      summary: Send a notification to all users (or one)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BroadcastBody" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Withdrawal:
      type: object
      properties:
        id: { type: integer }
        amount: { type: number }
        bankName: { type: string }
        accountNumber: { type: string }
        accountName: { type: string }
        status: { type: string, enum: [pending, approved, rejected, paid] }
        requestedAt: { type: string, format: date-time }
        decidedAt: { type: string, format: date-time, nullable: true }
        adminNote: { type: string, nullable: true }
      required: [id, amount, bankName, accountNumber, accountName, status, requestedAt]
    RequestWithdrawalBody:
      type: object
      properties:
        amount: { type: number }
      required: [amount]
...
    SupportTicket:
      type: object
      properties:
        id: { type: integer }
        subject: { type: string }
        status: { type: string, enum: [open, closed] }
        userEmail: { type: string, nullable: true }
        lastMessagePreview: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, subject, status, createdAt, updatedAt]
    SupportTicketDetail:
      type: object
      properties:
        ticket: { $ref: "#/components/schemas/SupportTicket" }
        messages:
          type: array
          items: { $ref: "#/components/schemas/SupportMessage" }
      required: [ticket, messages]
    SupportMessage:
      type: object
      properties:
        id: { type: integer }
        ticketId: { type: integer }
        senderRole: { type: string, enum: [user, admin] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, ticketId, senderRole, body, createdAt]
...
    PostTicketMessageBody:
      type: object
      properties:
        body: { type: string }
      required: [body]
...
    AdminUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        fullName: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        referralCode: { type: string }
        referenceCode: { type: string }
        status: { type: string }
        role: { type: string }
        balance: { type: number }
        createdAt: { type: string, format: date-time }
      required: [id, email, referralCode, referenceCode, status, role, balance, createdAt]
    AdminUpdateBalanceBody:
      type: object
      properties:
        delta: { type: number, description: "Positive to credit, negative to debit" }
        reason: { type: string }
      required: [delta, reason]
    AdminSetStatusBody:
      type: object
      properties:
        status: { type: string, enum: [pending, active, suspended] }
      required: [status]
    AdminPayment:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: string }
        amount: { type: number }
        senderName: { type: string }
        note: { type: string, nullable: true }
        narration: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, userId, userEmail, amount, senderName, narration, status, createdAt]
    AdminDecidePaymentBody:
      type: object
      properties:
        decision: { type: string, enum: [confirm, reject] }
        note: { type: string }
      required: [decision]
    AdminWithdrawal:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: string }
        amount: { type: number }
        bankName: { type: string }
        accountNumber: { type: string }
        accountName: { type: string }
        status: { type: string }
        requestedAt: { type: string, format: date-time }
      required: [id, userId, userEmail, amount, bankName, accountNumber, accountName, status, requestedAt]
    AdminDecideWithdrawalBody:
      type: object
      properties:
        decision: { type: string, enum: [approve, reject, paid] }
        note: { type: string }
      required: [decision]
    AffiliatePayout:
      type: object
      properties:
        id: { type: integer }
        referrerId: { type: integer }
        referrerEmail: { type: string }
        referredUserId: { type: integer }
        referredUserEmail: { type: string }
        amount: { type: number }
        status: { type: string, enum: [pending, paid] }
        createdAt: { type: string, format: date-time }
      required: [id, referrerId, referrerEmail, referredUserId, referredUserEmail, amount, status, createdAt]
    BroadcastBody:
      type: object
      properties:
        title: { type: string }
        body: { type: string }
        userId: { type: integer, nullable: true, description: "Target a specific user, omit for broadcast" }
      required: [title, body]