# lib/api-zod/src/generated/types/adminSetStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WINR platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSetStatusBodyStatus } from "./adminSetStatusBodyStatus";

export interface AdminSetStatusBody {
  status: AdminSetStatusBodyStatus;
}