# lib/api-zod/src/generated/types/adminWithdrawal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WINR platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminWithdrawal {
  id: number;
  userId: number;
  userEmail: string;
  amount: number;
  bankName: string;
  accountNumber: string;
  accountName: string;
  status: string;
  requestedAt: Date;
}