# lib/api-zod/src/generated/types/supportMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WINR platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SupportMessageSenderRole } from "./supportMessageSenderRole";

export interface SupportMessage {
  id: number;
  ticketId: number;
  senderRole: SupportMessageSenderRole;
  body: string;
  createdAt: Date;
}