# lib/api-zod/src/generated/types/supportTicketDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WINR platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SupportMessage } from "./supportMessage";
import type { SupportTicket } from "./supportTicket";

export interface SupportTicketDetail {
  ticket: SupportTicket;
  messages: SupportMessage[];
}